HBX-2835: Move classes from package 'org.hibernate.tool.orm.jbt.util' to package 'org.hibernate.tool.orm.jbt.internal.util'   - Move class 'org.hibernate.tool.orm.jbt.util.JpaConfiguration'   - Move test class 'org.hibernate.tool.orm.jbt.util.JpaConfigurationTest'
