+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/JpaConfiguration.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.orm.jbt.util;
+package org.hibernate.tool.orm.jbt.internal.util;
 
 import java.io.File;
 import java.util.Iterator;
@@ -16,8 +16,6 @@ import org.hibernate.mapping.Table;
 import org.hibernate.tool.api.reveng.RevengStrategy;
 import org.hibernate.tool.orm.jbt.api.wrp.SessionFactoryWrapper;
 import org.hibernate.tool.orm.jbt.internal.factory.SessionFactoryWrapperFactory;
-import org.hibernate.tool.orm.jbt.internal.util.ExtendedConfiguration;
-import org.hibernate.tool.orm.jbt.internal.util.HibernateToolsPersistenceProvider;
 import org.w3c.dom.Document;
 import org.xml.sax.EntityResolver;
